refactor(FR-1006): delete classic session launcher and use neo as default (#3690)
resolves #3676 (FR-1006)
<!-- replace NNN, MMM with the GitHub issue number and the corresponding Jira issue number. -->
<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

### Delete classic session launcher and use 'NEO' as default.

This PR is about deleting all legacy files against Classic Session Launcher and using Neo Session Launcher by default.
The associated settings feature is also removed.
**How to test:**
- You should check the behavior of existing components that are using `backend-ai-session-launcher.ts`.
   - When you click the start button in the legacy session list, check if the Neo session launcher shows correctly.
   - You should verify that all the actions on the Import & run page are working as they should.

**Checklist:** (if applicable)

- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after
