+++ resources/i18n/el.json
@@ -1803,7 +1803,6 @@
     "BootstrapScriptDescription": "Το σενάριο Bootstrap θα εκτελεστεί μόνο μία φορά μετά τη δημιουργία συνεδρίας",
     "BootstrapScriptEmpty": "Το σενάριο είναι κενό. \nΕισαγάγετε το σενάριο.",
     "BootstrapScriptUpdated": "Το σενάριο Bootstrap ενημερώθηκε.",
-    "ClassicSessionLauncher": "Classic Session Launcher",
     "ClearSSHKeypairInput": "Το γραπτό πληκτρολόγιο SSH θα αφαιρεθεί. Εισαι σιγουρος οτι θελεις να συνεχισεις?",
     "ConfigFilename": "Διαμόρφωση ονόματος αρχείου",
     "CurrentSSHPublicKey": "Τρέχον δημόσιο κλειδί:",
@@ -1812,7 +1811,6 @@
     "DescAutoLogout": "Αποσυνδεθείτε αυτόματα εάν όλες οι συνδεδεμένες σελίδες είναι κλειστές εκτός από τις σελίδες που προέρχονται από περίοδο σύνδεσης για εφαρμογές (π.χ. jupyter notebook, web terminal κ.λπ.)",
     "DescAutomaticUpdateCheck": "Ειδοποίηση όταν κυκλοφορήσει μια νέα έκδοση webui. <br /> Όταν το Διαδίκτυο είναι ασταθές ή εκτός σύνδεσης, αυτή η επιλογή θα απενεργοποιηθεί αυτόματα. Μπορείτε να αλλάξετε αυτόν τον διακόπτη για να ξεκινήσετε ξανά τον έλεγχο αυτόματης ενημέρωσης.",
     "DescBetaFeatures": "Χρησιμοποιήστε λειτουργίες beta για διεπαφή χρήστη Ιστού. <br /> Οι λειτουργίες beta ενδέχεται να είναι ασταθείς. Ορισμένες λειτουργίες beta ενδέχεται να μην υιοθετηθούν ως επίσημη λειτουργία.",
-    "DescClassicSessionLauncher": "Χρησιμοποιήστε το πρόγραμμα εκκίνησης συνεδρίας Classic αντί για το πρόγραμμα εκκίνησης συνεδρίας NEO. \nΤο πρόγραμμα εκκίνησης περιόδου λειτουργίας Classic ενδέχεται να μην έχει τις πιο πρόσφατες δυνατότητες.",
     "DescDesktopNotification": "Ενεργοποίηση ή απενεργοποίηση ειδοποίησης επιφάνειας εργασίας. <br /> Εάν είναι ενεργοποιημένο, το Backend.AI χρησιμοποιεί επίσης ενσωματωμένο σύστημα ειδοποιήσεων OS. Η απενεργοποίηση αυτής της επιλογής δεν επηρεάζει τις ειδοποιήσεις στο περιβάλλον χρήστη ιστού.",
     "DescExperimentalDashboard": "Χρησιμοποιήστε τη σελίδα ταμπλό αντί για συνοπτική σελίδα.",
     "DescKeepLoginSessionInformation": "Αφήστε την εφαρμογή webui να διατηρήσει τις τρέχουσες πληροφορίες σύνδεσης κατά την επόμενη φορά. <br /> Εάν η επιλογή είναι απενεργοποιημένη, οι πληροφορίες σύνδεσης θα εκκαθαρίζονται σε κάθε αποσύνδεση.",
