+++ resources/i18n/ko.json
@@ -1811,7 +1811,6 @@
     "BootstrapScriptDescription": "부트스트랩 스크립트는 세션 생성 후 한 번만 실행됩니다",
     "BootstrapScriptEmpty": "스크립트가 비어있습니다. 스크립트를 입력해 주세요.",
     "BootstrapScriptUpdated": "부트스트랩 스크립트가 갱신되었습니다.",
-    "ClassicSessionLauncher": "Classic 세션 런처",
     "ClearSSHKeypairInput": "지금까지 작성된 SSH 키페어는 삭제될 예정입니다. 계속하겠습니까?",
     "ConfigFilename": "구성 파일 이름 선택",
     "CurrentSSHPublicKey": "현재 공개 키: ",
@@ -1820,7 +1819,6 @@
     "DescAutoLogout": "세션 내 앱을 실행하기 위해 생성된 페이지를 제외한 모든 Backend.AI 웹 UI 페이지가 닫힐 경우 자동으로 로그아웃됩니다.",
     "DescAutomaticUpdateCheck": "새로운 웹 UI 버전이 출시된 경우 알려줍니다.<br />오프라인 또는 인터넷 연결 상태가 불안정한 환경에서는 몇 차례 시도후 자동으로 사용이 중지됩니다. 이 경우 이 스위치를 켜서 다시 자동 업데이트 체크를 시작할 수 있습니다.",
     "DescBetaFeatures": "GUI용으로 테스트중인 기능을 사용합니다.<br />베타 기능은 불안정할 수 있으며, 정식 버전에서는 채택되지 않을 수 있습니다.",
-    "DescClassicSessionLauncher": "최신의 NEO 세션 런처 대신 Classic 세션 런처를 사용합니다. Classic 세션 런처에는 최신 기능이 지원되지 않을 수 있습니다.",
     "DescDesktopNotification": "데스크탑 알림을 사용합니다.<br />이 옵션이 켜져 있을때, Backend.AI는 운영체제의 알림 기능을 함께 사용합니다. 이 옵션을 꺼도 웹 UI내의 알림은 계속 동작합니다.",
     "DescExperimentalDashboard": "요약 페이지 대신 대시보드 페이지를 사용합니다.",
     "DescKeepLoginSessionInformation": "웹 UI 앱이 다음 앱 사용시까지 로그인 세션 정보를 저장합니다.<br />이 옵션이 꺼져있을 경우, 매 종료시마다 로그인 정보가 자동으로 삭제됩니다.",
