+++ resources/i18n/mn.json
@@ -1804,7 +1804,6 @@
     "BootstrapScriptDescription": "Ачаалагч скрипт нь сесс үүсгэсний дараа зөвхөн нэг удаа хийгдэх болно",
     "BootstrapScriptEmpty": "Скрипт хоосон байна. \nСкриптийг оруулна уу.",
     "BootstrapScriptUpdated": "Ачаалагч скрипт шинэчлэгдсэн.",
-    "ClassicSessionLauncher": "Classic сесс эхлүүлэгч",
     "ClearSSHKeypairInput": "Бичсэн SSH товчлуурыг устгах болно. Үргэлжлүүлэхийг хүсч байна уу?",
     "ConfigFilename": "Файлын нэрийг тохируулах",
     "CurrentSSHPublicKey": "Одоогийн нийтийн түлхүүр:",
@@ -1813,7 +1812,6 @@
     "DescAutoLogout": "Аппликешнд зориулж суулгасан хуудсуудаас бусад нэвтэрсэн бүх хуудсыг хаасан тохиолдолд автоматаар гарах болно (ж.нь. Jupyter Notebook, вэб терминал гэх мэт).",
     "DescAutomaticUpdateCheck": "Webui-ийн шинэ хувилбар гарах үед мэдэгдэх. <br /> Интернэт тогтворгүй эсвэл офлайн байгаа тохиолдолд энэ тохируулга автоматаар унтарна. Автомат шинэчлэлтийн шалгалтыг дахин эхлүүлэхийн тулд та энэ шилжүүлэгчийг асааж болно.",
     "DescBetaFeatures": "Вэб UI-д бета функцийг ашиглах. <br /> бета функцууд тогтворгүй байж магадгүй. Зарим бета функцийг албан ёсны шинж чанар болгон ашиглахгүй байж магадгүй юм.",
-    "DescClassicSessionLauncher": "NEO сесс эхлүүлэгчийн оронд Classic сесс эхлүүлэгчийг ашиглана уу. \nClassic сесс эхлүүлэгч нь хамгийн сүүлийн үеийн функцгүй байж магадгүй.",
     "DescDesktopNotification": "Ширээний мэдэгдлийг асаах эсвэл унтраах. <br /> Хэрэв асаалттай бол Backend.AI үйлдлийн системд суурилуулсан мэдэгдлийн системийг ашигладаг. Энэ сонголтыг унтраавал вэб интерфэйс дээрх мэдэгдэлд нөлөөлөхгүй.",
     "DescExperimentalDashboard": "Хураангуй хуудасны оронд хяналтын самбар ашиглана уу.",
     "DescKeepLoginSessionInformation": "Webui програмыг нэвтрэх Sessionийн мэдээллийг дараагийн удаа хадгалахыг зөвшөөрнө үү. <br /> Хэрэв тохиргоог унтраасан бол нэвтрэх тухай мэдээллийг нэвтрэх бүрээс хасах болно.",
