+++ resources/i18n/pl.json
@@ -1807,7 +1807,6 @@
     "BootstrapScriptDescription": "Skrypt Bootstrap zostanie wykonany tylko raz po utworzeniu sesji.",
     "BootstrapScriptEmpty": "Skrypt jest pusty. \nProszę wprowadzić skrypt.",
     "BootstrapScriptUpdated": "Zaktualizowano skrypt Bootstrap.",
-    "ClassicSessionLauncher": "Classic program uruchamiający sesję",
     "ClearSSHKeypairInput": "Zapisana para kluczy SSH zostanie usunięta. Jesteś pewien, że chcesz kontynuować?",
     "ConfigFilename": "Nazwa pliku konfiguracyjnego",
     "CurrentSSHPublicKey": "Aktualny klucz publiczny:",
@@ -1816,7 +1815,6 @@
     "DescAutoLogout": "Wyloguj się automatycznie, jeśli wszystkie zalogowane strony są zamknięte, z wyjątkiem stron pochodzących z sesji dla aplikacji (np. notatnik jupyter, terminal internetowy itp.)",
     "DescAutomaticUpdateCheck": "Powiadom o wydaniu nowej wersji webui. <br /> W przypadku, gdy Internet jest niestabilny lub offline, ta opcja zostanie automatycznie wyłączona. Możesz przełączyć ten przełącznik, aby ponownie rozpocząć automatyczne sprawdzanie aktualizacji.",
     "DescBetaFeatures": "Korzystaj z funkcji beta interfejsu internetowego. <br /> Funkcje beta mogą być niestabilne. Niektóre funkcje beta mogą nie zostać przyjęte jako funkcja oficjalna.",
-    "DescClassicSessionLauncher": "Użyj Classic programu uruchamiającego sesję zamiast programu uruchamiającego sesję NEO. \nClassic program uruchamiający sesję może nie mieć najnowszych funkcji.",
     "DescDesktopNotification": "Włącz lub wyłącz powiadomienia na pulpicie. <br /> Jeśli jest włączony, Backend.AI używa również wbudowanego systemu powiadomień. Wyłączenie tej opcji nie wpływa na powiadomienia w interfejsie internetowym.",
     "DescExperimentalDashboard": "Użyj strony Dashboard zamiast strony podsumowującej.",
     "DescKeepLoginSessionInformation": "Pozwól aplikacji webui zachować bieżące informacje o sesji logowania następnym razem. <br /> Jeśli opcja jest wyłączona, dane logowania będą usuwane przy każdym wylogowaniu.",
