+++ resources/i18n/ru.json
@@ -1807,7 +1807,6 @@
     "BootstrapScriptDescription": "Bootstrap-скрипт будет выполняться только один раз после создания сессии",
     "BootstrapScriptEmpty": "Скрипт пуст. \nПожалуйста, введите сценарий.",
     "BootstrapScriptUpdated": "Обновлен скрипт начальной загрузки.",
-    "ClassicSessionLauncher": "__NOT_TRANSLATED__",
     "ClearSSHKeypairInput": "Записанная пара ключей SSH будет удалена. Вы уверены что хотите продолжить?",
     "ConfigFilename": "Имя файла конфигурации",
     "CurrentSSHPublicKey": "Текущий открытый ключ:",
@@ -1816,7 +1815,6 @@
     "DescAutoLogout": "Автоматический выход из системы, если все зарегистрированные страницы закрыты, кроме страниц, созданных в сеансе для приложений (например, блокнот jupyter, веб-терминал и т. Д.)",
     "DescAutomaticUpdateCheck": "Уведомлять о выходе новой версии webui. <br /> Если Интернет нестабилен или отключен, эта опция будет автоматически отключена. Вы можете переключить этот переключатель, чтобы снова запустить автоматическую проверку обновлений.",
     "DescBetaFeatures": "Используйте бета-функции для веб-интерфейса. <br /> Бета-функции могут быть нестабильными. Некоторые бета-функции могут быть не приняты в качестве официальных.",
-    "DescClassicSessionLauncher": "__NOT_TRANSLATED__",
     "DescDesktopNotification": "Включение или отключение уведомлений на рабочем столе. <br /> Если включено, Backend.AI также использует встроенную систему уведомлений ОС. Отключение этого параметра не влияет на уведомления в веб-интерфейсе.",
     "DescExperimentalDashboard": "Используйте страницу панели приборной панели вместо сводной страницы.",
     "DescKeepLoginSessionInformation": "Разрешите приложению webui в следующий раз сохранить текущую информацию о сеансе входа. <br /> Если опция выключена, информация для входа в систему будет очищаться при каждом выходе из системы.",
