+++ resources/i18n/th.json
@@ -1788,7 +1788,6 @@
     "BootstrapScriptDescription": "สคริปต์บูตสแตรปจะถูกดำเนินการเพียงครั้งเดียวหลังจากสร้างเซสชัน",
     "BootstrapScriptEmpty": "สคริปต์ว่างเปล่า \nกรุณาป้อนสคริปต์",
     "BootstrapScriptUpdated": "อัปเดตสคริปต์บูตสแตรปแล้ว",
-    "ClassicSessionLauncher": "ตัวเปิดเซสชัน Classic",
     "ClearSSHKeypairInput": "คู่คีย์ SSH ที่เขียนไว้จะถูกลบ คุณแน่ใจหรือไม่ที่จะดำเนินการต่อ?",
     "ConfigFilename": "ชื่อไฟล์การกำหนดค่า",
     "CurrentSSHPublicKey": "คีย์สาธารณะปัจจุบัน: ",
@@ -1797,7 +1796,6 @@
     "DescAutoLogout": "ออกจากระบบโดยอัตโนมัติหากปิดหน้าที่เข้าสู่ระบบทั้งหมดยกเว้นหน้าที่มาจากเซสชันสำหรับแอป (เช่น jupyter notebook, เทอร์มินัลเว็บ ฯลฯ)",
     "DescAutomaticUpdateCheck": "แจ้งเตือนเมื่อมีการปล่อยเวอร์ชันใหม่ของส่วนติดต่อผู้ใช้บนเว็บ<br />ในที่ที่อินเทอร์เน็ตไม่เสถียรหรือออฟไลน์ ตัวเลือกนี้จะถูกปิดโดยอัตโนมัติ คุณสามารถสลับสวิตช์นี้เพื่อเริ่มการตรวจสอบการอัปเดตอัตโนมัติอีกครั้ง",
     "DescBetaFeatures": "ใช้คุณสมบัติเบต้าสำหรับส่วนติดต่อผู้ใช้บนเว็บ<br />คุณสมบัติเบต้าอาจไม่เสถียร บางคุณสมบัติเบต้าอาจไม่ได้รับการยอมรับให้เป็นคุณสมบัติอย่างเป็นทางการ",
-    "DescClassicSessionLauncher": "ใช้ตัวเรียกใช้เซสชัน Classic แทนตัวเรียกใช้เซสชัน NEO \nตัวเรียกใช้เซสชัน Classic อาจไม่มีคุณสมบัติล่าสุด",
     "DescDesktopNotification": "เปิดหรือปิดการแจ้งเตือนบนเดสก์ท็อป <br />หากเปิด Backend.AI จะใช้ระบบการแจ้งเตือนในตัวของระบบปฏิบัติการด้วย การปิดตัวเลือกนี้จะไม่ส่งผลกระทบต่อการแจ้งเตือนภายในส่วนติดต่อผู้ใช้บนเว็บ",
     "DescExperimentalDashboard": "ใช้หน้าแดชบอร์ดแทนหน้าสรุป",
     "DescKeepLoginSessionInformation": "ให้แอป webui เก็บข้อมูลเซสชันเข้าสู่ระบบปัจจุบันไว้สำหรับครั้งต่อไป<br />หากปิดตัวเลือกนี้ ข้อมูลการเข้าสู่ระบบจะถูกล้างทุกครั้งที่ออกจากระบบ",
