+++ resources/i18n/tr.json
@@ -1806,7 +1806,6 @@
     "BootstrapScriptDescription": "Bootstrap komut dosyası oturum oluşturulduktan sonra yalnızca bir kez çalıştırılacaktır",
     "BootstrapScriptEmpty": "Senaryo boş. \nLütfen betiği girin.",
     "BootstrapScriptUpdated": "Önyükleme komut dosyası güncellendi.",
-    "ClassicSessionLauncher": "Classic Oturum Başlatıcı",
     "ClearSSHKeypairInput": "Yazılı SSH Anahtar Çifti kaldırılacaktır. Devam etmek istediğine emin misin?",
     "ConfigFilename": "Yapılandırma dosyası adı",
     "CurrentSSHPublicKey": "Geçerli ortak anahtar:",
@@ -1815,7 +1814,6 @@
     "DescAutoLogout": "Uygulamalar için oturum kaynaklı sayfalar (örneğin jupyter notebook, web terminal, vb.)",
     "DescAutomaticUpdateCheck": "Yeni bir webui sürümü yayınlandığında bildir. <br /> İnternetin kararsız veya çevrimdışı olduğu durumlarda bu seçenek otomatik olarak kapatılacaktır. Otomatik güncelleme kontrolünü yeniden başlatmak için bu anahtarı değiştirebilirsiniz.",
     "DescBetaFeatures": "Web kullanıcı arayüzü için beta özelliklerini kullanın. <br /> Beta özellikleri kararsız olabilir. Bazı beta özellikler resmi özellik olarak kabul edilmeyebilir.",
-    "DescClassicSessionLauncher": "NEO oturum başlatıcısı yerine Classic oturum başlatıcıyı kullanın. \nClassic oturum başlatıcı en son özelliklere sahip olmayabilir.",
     "DescDesktopNotification": "Masaüstü bildirimini açın veya kapatın. <br /> Açıksa, Backend.AI, OS yerleşik bildirim sistemini de kullanır. Bu seçeneği kapatmak, web kullanıcı arayüzündeki bildirimleri etkilemez.",
     "DescExperimentalDashboard": "Özet sayfası yerine gösterge paneli sayfasını kullanın.",
     "DescKeepLoginSessionInformation": "Webui uygulamasının bir dahaki sefere mevcut oturum açma bilgilerini tutmasına izin verin. <br /> Seçenek kapatılırsa, oturum açma bilgileri her oturum kapatıldığında temizlenir.",
