+++ resources/i18n/vi.json
@@ -1807,7 +1807,6 @@
     "BootstrapScriptDescription": "Tập lệnh Bootstrap sẽ chỉ được thực thi một lần sau khi tạo phiên",
     "BootstrapScriptEmpty": "Kịch bản trống rỗng. \nVui lòng nhập kịch bản.",
     "BootstrapScriptUpdated": "Đã cập nhật tập lệnh Bootstrap.",
-    "ClassicSessionLauncher": "Trình khởi chạy phiên Classic",
     "ClearSSHKeypairInput": "Cặp khóa SSH đã viết sẽ bị xóa. Bạn có chắc chắn muốn tiếp tục không?",
     "ConfigFilename": "Định cấu hình tên tệp",
     "CurrentSSHPublicKey": "Khóa công khai hiện tại:",
@@ -1816,7 +1815,6 @@
     "DescAutoLogout": "Tự động đăng xuất nếu tất cả các trang đã đăng nhập đều bị đóng ngoại trừ các trang gốc phiên cho ứng dụng (ví dụ: sổ ghi chép jupyter, thiết bị đầu cuối web, v.v.)",
     "DescAutomaticUpdateCheck": "Thông báo khi phiên bản webui mới được phát hành. <br /> Trong trường hợp Internet không ổn định hoặc ngoại tuyến, tùy chọn này sẽ tự động bị tắt. Bạn có thể bật công tắc này để bắt đầu lại quá trình kiểm tra cập nhật tự động.",
     "DescBetaFeatures": "Sử dụng các tính năng beta cho giao diện người dùng Web. <br /> Các tính năng beta có thể không ổn định. Một số tính năng beta có thể không được chấp nhận làm tính năng chính thức.",
-    "DescClassicSessionLauncher": "Sử dụng trình khởi chạy phiên Classic thay vì trình khởi chạy phiên NEO. \nTrình khởi chạy phiên Classic có thể không có các tính năng mới nhất.",
     "DescDesktopNotification": "Bật hoặc tắt thông báo trên màn hình. <br /> Nếu được bật, Backend.AI cũng sử dụng hệ thống thông báo tích hợp sẵn của OS. Tắt tùy chọn này không ảnh hưởng đến các thông báo trong giao diện người dùng web.",
     "DescExperimentalDashboard": "Sử dụng trang bảng điều khiển thay vì trang tóm tắt.",
     "DescKeepLoginSessionInformation": "Hãy để ứng dụng webui giữ thông tin phiên đăng nhập hiện tại vào lần sau. <br /> Nếu tùy chọn bị tắt, thông tin đăng nhập sẽ bị xóa mỗi lần đăng xuất.",
