+++ resources/i18n/zh-CN.json
@@ -1807,7 +1807,6 @@
     "BootstrapScriptDescription": "Bootstrap 脚本只会在会话创建后执行一次",
     "BootstrapScriptEmpty": "脚本是空的。\n请输入脚本。",
     "BootstrapScriptUpdated": "引导脚本已更新。",
-    "ClassicSessionLauncher": "Classic会话启动器",
     "ClearSSHKeypairInput": "书面 SSH 密钥对将被删除。你确定你要继续吗？",
     "ConfigFilename": "配置文件名",
     "CurrentSSHPublicKey": "当前公钥：",
@@ -1816,7 +1815,6 @@
     "DescAutoLogout": "如果所有登录页面都关闭，除了应用程序的会话发起页面（例如 jupyter 笔记本、网络终端等），则自动注销",
     "DescAutomaticUpdateCheck": "发布新的 webui 版本时通知。 <br /> 在互联网不稳定或离线的情况下，此选项将自动关闭。您可以切换此开关以再次启动自动更新检查。",
     "DescBetaFeatures": "使用 Web UI 的测试版功能。 <br /> Beta 功能可能不稳定。某些测试版功能可能不会被采用为正式功能。",
-    "DescClassicSessionLauncher": "使用 Classic 会话启动器而不是 NEO 会话启动器。 \nClassic 会话启动器可能不具有最新功能。",
     "DescDesktopNotification": "打开或关闭桌面通知。 <br /> 如果开启，Backend.AI 也会使用操作系统内置的通知系统。关闭此选项不会影响 Web UI 中的通知。",
     "DescExperimentalDashboard": "使用仪表板页面而不是摘要页面。",
     "DescKeepLoginSessionInformation": "下次让 webui 应用程序保留当前登录会话信息。 <br /> 如果关闭该选项，则每次注销时都会清除登录信息。",
