+++ resources/i18n/zh-TW.json
@@ -1806,7 +1806,6 @@
     "BootstrapScriptDescription": "Bootstrap 脚本只会在会话创建后执行一次",
     "BootstrapScriptEmpty": "腳本是空的。\n請輸入腳本。",
     "BootstrapScriptUpdated": "引導腳本已更新。",
-    "ClassicSessionLauncher": "Classic會話啟動器",
     "ClearSSHKeypairInput": "書面 SSH 密鑰對將被刪除。你確定你要繼續嗎？",
     "ConfigFilename": "配置文件名",
     "CurrentSSHPublicKey": "當前公鑰：",
@@ -1815,7 +1814,6 @@
     "DescAutoLogout": "如果所有登錄頁面都關閉，除了應用程序的會話發起頁面（例如 jupyter 筆記本、網絡終端等），則自動註銷",
     "DescAutomaticUpdateCheck": "發布新的 webui 版本時通知。 <br /> 在互聯網不穩定或離線的情況下，此選項將自動關閉。您可以切換此開關以再次啟動自動更新檢查。",
     "DescBetaFeatures": "使用 Web UI 的測試版功能。 <br /> Beta 功能可能不穩定。某些測試版功能可能不會被採用為正式功能。",
-    "DescClassicSessionLauncher": "使用 Classic 會話啟動器而不是 NEO 會話啟動器。 \nClassic 會話啟動器可能不具有最新功能。",
     "DescDesktopNotification": "打開或關閉桌面通知。 <br /> 如果開啟，Backend.AI 也會使用操作系統內置的通知系統。關閉此選項不會影響 Web UI 中的通知。",
     "DescExperimentalDashboard": "使用儀表板頁面而不是摘要頁面。",
     "DescKeepLoginSessionInformation": "下次讓 webui 應用程序保留當前登錄會話信息。 <br /> 如果關閉該選項，則每次註銷時都會清除登錄信息。",
